+++ buildSrc/src/main/kotlin/dev/muon/medieval/gradle/Versions.kt
@@ -1,7 +1,7 @@
 package dev.muon.medieval.gradle
 
 object Versions {
-    const val MOD = "1.4.0"
+    const val MOD = "1.5.0"
 
     const val MINECRAFT = "1.21.1"
     const val PARCHMENT_MINECRAFT = "1.21"
@@ -25,7 +25,7 @@ object Versions {
     const val JEI = "19.17.0.193"
     const val EMI = "1.1.13+1.21.1"
     const val JAVA = 21
-    const val CLOTH_CONFIG_VERSION = "15.0.130"
+    const val CLOTH_CONFIG_VERSION = "15.0.140"
 
     const val GECKOLIB = "1.21.1:4.7.3"
     const val PLAYER_ANIMATOR = "2.0.1+1.21.1"