+++ neoforge/src/main/java/dev/muon/medieval/platform/MedievalPlatformHelperNeoForge.java
@@ -2,6 +2,8 @@ package dev.muon.medieval.platform;
 
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.loading.FMLLoader;
+import net.neoforged.fml.loading.LoadingModList;
+import net.neoforged.fml.loading.moddiscovery.ModInfo;
 
 public class MedievalPlatformHelperNeoForge implements MedievalPlatformHelper {
 
@@ -12,6 +14,9 @@ public class MedievalPlatformHelperNeoForge implements MedievalPlatformHelper {
 
     @Override
     public boolean isModLoaded(String modId) {
+        if (ModList.get() == null) {
+        return LoadingModList.get().getMods().stream().map(ModInfo::getModId).anyMatch(modId::equals);
+    }
         return ModList.get().isLoaded(modId);
     }